Chart fixes for ngx-chart responsiveness bugs.
Chart was resizing out of control, but there were some issues with attempting to set a max-height or expecting the chart to behave normally when the screen was resized.

This commit contains a few workarounds to mitigate those issues. See the following issues:

swimlane/ngx-charts#1248
swimlane/ngx-charts#783
swimlane/ngx-charts#650

as well as

https://stackoverflow.com/questions/48423539/ngxcharts-resizing-on-data-change