Add more robust client side error handling.
Used Angular 7 best practices to implement error handling on the client side.
Currently error handler just alerts in the browser, but next step will be to display a popup.

Also removed some duplicate code on the server side, and made errors not include the exception message since the messages will be displayed to the user.